:art: move "generate_model_family" to converter module
