+++ modelci/hub/converter/__init__.py
@@ -15,10 +15,11 @@
 #  permissions and limitations under the License.
 
 from .converter import convert
+from .converter import generate_model_family
 from .to_trt import TRTConverter
 from .to_tfs import TFSConverter
 from .to_onnx import ONNXConverter
 from .to_pytorch import PyTorchConverter
 from .to_torchscript import TorchScriptConverter
 
-__all__ = ["convert", "TRTConverter", "TFSConverter", "ONNXConverter", "PyTorchConverter", "TorchScriptConverter"]
+__all__ = ["convert", "generate_model_family", "TRTConverter", "TFSConverter", "ONNXConverter", "PyTorchConverter", "TorchScriptConverter"]
